refactor(transfermarkt): split LiveMatchParser into focused parser classes
- Extract CompetitionParser for competition info extraction
- Extract MatchRowParser for live table row parsing
- Extract MatchResultParser for match result/status parsing
- Create thin LiveMatchParser facade orchestrating the above

Fixes CLAUDE.md convention: 357 LOC -> 69 LOC (under 150 max)
